+++ test/test279.js
@@ -15,7 +15,8 @@ describe('Test 279 IE9 tests', function() {
   it('1. Detect if it is IE9', function(done) {
     var data = [{city:"London",population:5000000}, {city:"Moscow",population:12000000}, ];
 //    alasql('SELECT * INTO TSV("aaa.txt",{headers:true}) FROM ?',[data]);
-    alasql('SELECT * INTO XLS("aaa.xls",{headers:true}) FROM ?',[data]);
+//    alasql('SELECT * INTO XLS("aaa.xls",{headers:true}) FROM ?',[data]);
+    alasql('SELECT * INTO XLSXML("aaa.xls",{headers:true}) FROM ?',[data]);
     done();
   });