+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * SEARCH objects
 * SEARCH graph
 * "name" as name for graph vertices and edges
+* Added INSTANCEOF selector
+* Added CLASS selector
+* * selector, + selector, ? selector, !selector for SEARCH in JSON and graphs
+* XML() from function
 
 ### 0.1.0 (aka 0.0.52) "Venice" (02.05.2015 - 03.05.2015)
 * Added INFORMATION_SCHEMA from variable